Combine call initiator and call acceptor classes
