+++ conclave-enclave/src/main/kotlin/com/r3/conclave/enclave/internal/NativeHostCallInterface.kt
@@ -13,7 +13,7 @@ typealias StackFrame = CallInterfaceStackFrame<HostCallType>
  * This class is the implementation of the [HostCallInterface] for native enclaves.
  * It has three jobs:
  *  - Serve as the endpoint for calls to make to the host, see [com.r3.conclave.common.internal.CallInitiator]
- *  - Route calls from the host to the appropriate enclave side call handler, see [com.r3.conclave.common.internal.CallAcceptor]
+ *  - Route calls from the host to the appropriate enclave side call handler, see [com.r3.conclave.common.internal.CallInterface]
  *  - Handle the low-level details of the messaging protocol (ecalls and ocalls).
  */
 class NativeHostCallInterface : HostCallInterface() {
