+++ conclave-host/src/main/kotlin/com/r3/conclave/host/internal/NativeEnclaveCallInterface.kt
@@ -11,7 +11,7 @@ typealias StackFrame = CallInterfaceStackFrame<EnclaveCallType>
  * This class is the implementation of the [EnclaveCallInterface] for native enclaves.
  * It has three jobs:
  *  - Serve as the endpoint for calls to make to the enclave, see [com.r3.conclave.common.internal.CallInitiator]
- *  - Route calls from the enclave to the appropriate host side call handler, see [com.r3.conclave.common.internal.CallAcceptor]
+ *  - Route calls from the enclave to the appropriate host side call handler, see [com.r3.conclave.common.internal.CallInterface]
  *  - Handle the low-level details of the messaging protocol (ecalls and ocalls).
  */
 class NativeEnclaveCallInterface(private val enclaveId: Long) : EnclaveCallInterface() {
